feat(pluginCustomTransform): Impl. of custom transformation in form of a plugin
