+++ playground/custom_transform/src/index.ts
@@ -1,0 +1,1 @@
+document.querySelector('#root').innerHTML = 'Custom transformed';
