--- playground/custom_transformer/src/index.ts
@@ -1,1 +1,0 @@
-console.log('I am custom transformed and my home directory is: ', process.env.HOME);
