+++ src/index.ts
@@ -7,4 +7,5 @@ export { pluginPostCSS } from './plugins/core/plugin_postcss';
 export { pluginReplace } from './plugins/core/plugin_replace';
 export { pluginSass } from './plugins/core/plugin_sass';
 export { pluginStylus } from './plugins/core/plugin_stylus';
+export { pluginCustomTransform } from './plugins/core/plugin_customtransform';
 export { sparky } from './sparky/sparky';
