+++ src/plugins/core/plugin_customtransform.ts
@@ -1,0 +1,8 @@
+import { Context } from '../../core/Context';
+import { CustomTransformers } from 'typescript';
+
+export function pluginCustomTransform(customTransformers: CustomTransformers) {
+  return (ctx: Context) => {
+    ctx.customTransformers = customTransformers;
+  };
+}
