+++ src/production/stages/transpileStage.ts
@@ -1,15 +1,15 @@
 import { IProductionFlow } from '../main';
 import { ProductionModule } from '../ProductionModule';
 import * as ts from 'typescript';
-import { isRequireCall } from '../../transform/tsTransform';
 import { fixModuleSourceMap } from '../../sourcemaps/helpers';
+import { moduleTransformer } from '../../module-transformer/production';
 
-interface ITranspileStageProps {
+export interface ITranspileStageProps {
   flow: IProductionFlow;
   amount: number;
 }
 
-function addModule2ProductionSchema(props: ITranspileStageProps, pm: ProductionModule) {
+export function addModule2ProductionSchema(props: ITranspileStageProps, pm: ProductionModule) {
   const schema = props.flow.productionContext.schema;
   if (schema.indexOf(pm) > -1) {
     return;
@@ -19,68 +19,6 @@ function addModule2ProductionSchema(props: ITranspileStageProps, pm: ProductionM
     props.amount++;
     transpileProductionModule(props, pm);
   }
-}
-
-function moduleTransformer<T extends ts.Node>(
-  props: ITranspileStageProps,
-  pm: ProductionModule,
-): ts.TransformerFactory<T> {
-  const log = props.flow.ctx.log;
-  const config = props.flow.ctx.config;
-  return context => {
-    const webWorkers = pm.module.fastAnalysis && pm.module.fastAnalysis.workers;
-    const visit: ts.Visitor = node => {
-      if (webWorkers && ts.isNewExpression(node)) {
-        const newText = node.expression.getText();
-        if (newText === 'Worker' || newText === 'SharedWorker') {
-          if (node.arguments.length === 1) {
-            const firstArg = node.arguments[0];
-            const statement = firstArg['text'];
-            // map item
-            const item = webWorkers.find(item => {
-              return item.path === statement && item.type === newText;
-            });
-            if (item) {
-              return ts.createNew(ts.createIdentifier(newText), undefined, [ts.createStringLiteral(item.bundlePath)]);
-            }
-          }
-        }
-      }
-      if (isRequireCall(node)) {
-        const text = node['arguments'][0].text;
-
-        const target = pm.findDependantModule(text);
-        if (!target) {
-          if (config.target !== 'electron' && config.target !== 'server') {
-            if (!props.flow.ctx.config.shoudIgnorePackage(text)) {
-              if (log.props.ignoreStatementErrors && log.props.ignoreStatementErrors.includes(text)) {
-                return;
-              }
-              log.error('Problem when resolving require "$text" in $file', {
-                text: text,
-                file: pm.module.getShortPath(),
-              });
-            }
-          }
-        } else {
-          log.progressFormat(
-            target.module.isExecutable() ? 'Transpile' : 'Register',
-            `"${text}" from ${pm.module.getShortPath()} to ${target.module.getShortPath()}`,
-          );
-
-          addModule2ProductionSchema(props, target);
-          return ts.createCall(
-            ts.createPropertyAccess(ts.createIdentifier('$fsx'), ts.createIdentifier('r')),
-            undefined,
-            [ts.createNumericLiteral(target.getId().toString())],
-          );
-        }
-      }
-      return ts.visitEachChild(node, child => visit(child), context);
-    };
-
-    return node => ts.visitNode(node, visit);
-  };
 }
 
 function transpileProductionModule(props: ITranspileStageProps, prodModule: ProductionModule) {
@@ -95,24 +33,15 @@ function transpileProductionModule(props: ITranspileStageProps, prodModule: Prod
     compilerOptions.inlineSources = true;
   }
 
+  const customTransformers = ctx.customTransformers || {};
+
   const result = ts.transpileModule(text, {
     fileName: prodModule.module.props.absPath,
     compilerOptions: compilerOptions,
-
     transformers: {
-
-      // merge in any custom transformers (user-provided)
-      ...ctx.customTransformers,
-
-      // 2nd-level merge in transformers
-      after: [
-
-        // make sure core transformers are applied and executed first
-        moduleTransformer(props, prodModule),
-
-        // user-provided transformers
-        ...ctx.customTransformers.after,
-      ]
+      before: customTransformers.before || [],
+      after: [].concat(moduleTransformer(props, prodModule), customTransformers.after || []),
+      afterDeclarations: customTransformers.afterDeclarations || [],
     },
   });
